19th "CRUD messages,rooms,and user auth is done"

+++ base/views.py
@@ -1,11 +1,59 @@
 from django.shortcuts import render ,redirect
+from django.contrib import messages
+from django.http import HttpResponse
 from django.db.models import Q
-from .models import Room,Topic
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate,login,logout
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.decorators import login_required
+from .models import Room,Topic,Message
 from .forms import RoomForm
 
 def login_register_Page(request):
-    context ={}
-    return render (request,'base/login_register.html',context)    
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request,'User does not exist !')
+            
+        user = authenticate(request,username=username,password=password)
+        
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,"username OR password does not exist")
+    context ={'page':page}
+    return render (request,'base/login_register.html',context) 
+   
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerUser(request):
+    # page = 'register'
+    form = UserCreationForm()
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'Registration Failed!!!')
+            
+    return render (request,'base/login_register.html',{'form':form})
+
 def home(request):
     
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -25,9 +73,21 @@ def home(request):
 
 def room(request,id):
     room = Room.objects.get(id=id)
-    context = {'room':room}
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+    if request.method =='POST':
+        message = Message.objects.create(
+            user = request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room',id=room.id)
+    
+    context = {'room':room,'room_messages':room_messages,'participants':participants}
     return render (request, 'base/room.html',context)
 
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -39,9 +99,12 @@ def createRoom(request):
     context = {'form':form}
     return render (request,'base/room_form.html',context)
 
+@login_required(login_url='/login')
 def updateRoom(request,id):
     room = Room.objects.get(id=id)
     form = RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse("You can't make changes here !")
     if request.method == 'POST':
         form =RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -51,9 +114,33 @@ def updateRoom(request,id):
     context ={'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='/login')
+def updateMessage(request,id):
+    message = Message.objects.get(id=id)
+    if request.user != message.user:
+        return HttpResponse("You can't make changes here !")
+    if request.method == 'POST':
+        message.body = request.POST['body']
+        message.save()
+        return redirect ('home')
+    return render(request, 'base/update_message.html', {'message': message})
+
+@login_required(login_url='/login')
 def deleteRoom(request,id):
     room = Room.objects.get(id=id)
+    if request.user != room.host:
+        return HttpResponse("You can't delete this room!")
     if request.method == 'POST':
         room.delete()
         return redirect ('home')
     return render(request,'base/delete.html',{'obj':room})
+
+@login_required(login_url='/login')
+def deleteMessage(request,id):
+    message =Message.objects.get(id=id)
+    if request.user != message.user:
+        return HttpResponse("You can't delete this message!")
+    if request.method == 'POST':
+        message.delete()
+        return redirect ('home')
+    return render(request,'base/delete.html',{'obj':message})
